+++ lib/dhttpd.dart
@@ -1,33 +1,21 @@
-library dhttpd;
-
 import 'dart:async';
 import 'dart:io';
 
 import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart' as io;
 import 'package:shelf_static/shelf_static.dart';
-import 'package:shelf_cors/shelf_cors.dart' as shelf_cors;
 
 const int DEFAULT_PORT = 8080;
 const String DEFAULT_HOST = 'localhost';
 
 class Dhttpd {
   static Future<Dhttpd> start(
-      {String path,
-      int port: DEFAULT_PORT,
-      String allowOrigin,
-      address: DEFAULT_HOST}) async {
+      {String path, int port: DEFAULT_PORT, address: DEFAULT_HOST}) async {
     if (path == null) path = Directory.current.path;
 
     final handler = createStaticHandler(path, defaultDocument: 'index.html');
 
     final pipeline = const Pipeline().addMiddleware(logRequests());
-
-    if (allowOrigin != null) {
-      final corsHeaders = {'Access-Control-Allow-Origin': allowOrigin,};
-      pipeline.addMiddleware(
-          shelf_cors.createCorsHeadersMiddleware(corsHeaders: corsHeaders));
-    }
 
     HttpServer server =
         await io.serve(pipeline.addHandler(handler), address, port);